ThreatTrace/backend/requirements.txt
Flask
Flask-Cors
Flask-SocketIO
Flask-JWT-Extended
pymongo
python-dotenv
pandas
scikit-learn
requests
reportlab
bcrypt
